tests/requirements.txt
pytest==8.3.5
pytest-xdist==3.6.1
allure-pytest==2.13.5
requests==2.32.3
python-dotenv==1.0.1
